backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.3.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
